fix: Swift/MapLibre API corrections - style from temp file, MLN match expression, MLNShapeSource shape updates, Kotlin interop conversions, iOS 16 target

+++ iosApp/iosApp/Map/GeoJsonToFeatures.swift
@@ -53,7 +53,7 @@ enum GeoJsonToFeatures {
         guard shellPoints.count >= 3 else { return [] }
 
         let polygon = shellPoints.withUnsafeBufferPointer { buffer -> MLNPolygonFeature in
-            let feature = MLNPolygonFeature(coordinates: buffer.baseAddress!, count: buffer.count)
+            let feature = MLNPolygonFeature(coordinates: buffer.baseAddress!, count: UInt(buffer.count))
             feature.attributes = properties
             return feature
         }
